requirements.txt
pytest==4.3.1
Scrapy==1.6.0
sujmarkov==1.0.1
yapf==0.26.0
unicode-slugify==0.1.3
